Fixed post bug, added features
Fixed the issue of hanging when the posts arrray was 0.

Added moderator post styling.

User cannot unfollow moderator.

Added moderator flag to model.

Added dropdown for simulating posts as pagination.

Added moderator posts to the dashboard view.

+++ routes/account.js
@@ -59,6 +59,7 @@ router.get('/:short_id/view', isLoggedIn, async function(req, res, next){
   var userPosts = await getAllUserPosts(req.params.short_id)
 
   res.render('accountView', 
+  // DEV NOTE: hardcoding setting user posts to the default 25
   {
     userFound: userForView, 
     currentUser: req.user, 
@@ -79,13 +80,16 @@ router.post('/:short_id/view/more', isLoggedIn, async function(req, res, next){
   }
 
   var userPosts = await getAllUserPosts(req.params.short_id)
+  if(!req.body.postCounter == 'all') {
+    userPosts.slice(userPosts.length - req.body.postCounter)
+  }
 
   res.render('accountView', 
   {
     userFound: userForView, 
     currentUser: req.user, 
     userFollows, 
-    userPosts: userPosts.slice(userPosts.length - req.body.postCounter),
+    userPosts,
     moment,
     postCounter: req.body.postCounter,
   })
@@ -94,16 +98,18 @@ router.post('/:short_id/view/more', isLoggedIn, async function(req, res, next){
 
 router.post('/search', isLoggedIn, async function(req, res) {
 
-  console.log('search short id: ' + req.body.short_id)
+  // console.log('search short id: ' + req.body.short_id)
 
-  var userForView = await findUserForView(req.body.short_id);
+  // var userForView = await findUserForView(req.body.short_id);
 
-  if(userForView.short_id !== req.user.short_id) {
-    var userFollows = await doesUserFollow(req.user.short_id, req.body.short_id);
-    console.log('doesUserFollow function returns: ' + userFollows)
-  }
+  // if(userForView.short_id !== req.user.short_id) {
+  //   var userFollows = await doesUserFollow(req.user.short_id, req.body.short_id);
+  // }
 
-  res.render('accountView', {userFound: userForView, currentUser: req.user, userFollows})
+  // res.render('accountView', {userFound: userForView, currentUser: req.user, userFollows, postCounter: 25})
+
+  res.redirect('/account/' + req.body.short_id  + '/view');
+
 })
 
 router.post('/:followerShortId/:followingShortId/follow', isLoggedIn, function(req, res) {
@@ -123,23 +129,23 @@ router.post('/:followerShortId/:followingShortId/follow', isLoggedIn, function(r
 })
 
 router.post('/:followerShortId/:followingShortId/unfollow', isLoggedIn, function(req, res) {
-  console.log('unfollow clicked')
 
   var followingShortId = req.params.followingShortId;
   var followerShortId = req.params.followerShortId;
 
-  console.log('followingShortId: ' + followingShortId)
-  console.log('followerShortId: ' + followerShortId)
+  // DEV NOTE: make sure the the user is not entering the moderator in route
+  if(followingShortId != 'vHN_B6IA7') {
+    Follow.findOneAndDelete({'followerShortId': followerShortId, 'followingShortId': followingShortId}, (err, userFound) => {
+      if(err) {console.log(err)}
+      if(!userFound) {
+        console.log('no user found')
+      }
 
-  Follow.findOneAndDelete({'followerShortId': followerShortId, 'followingShortId': followingShortId}, (err, userFound) => {
-    if(err) {console.log(err)}
-    if(!userFound) {
-      console.log('no user found')
-    }
-   
-    console.log('the user found was: ' + userFound)
-    res.redirect('/account/followers')
-  })
+      res.redirect('/account/followers')
+    })
+  } else {
+    res.redirect('/dashboard/')
+  }
 })
 
 module.exports = router;
